Made note visible as default, fixed some small issues with margins, added clear all notes and clear notes for one page and also implemented export all notes for all pages

+++ src/js/lib/backbone.sync.amd.js
@@ -57,13 +57,17 @@ define([
 				console.log(e);
 			});
 		} else if (method == 'delete') {
-			req = db.remove('notes', model.url);
-			req.done(function () {
-				//sendResponse({msg: 'saved'});
-			});
-			req.fail(function (e) {
-				console.log(e);
-			});	
+			if(model == null) {
+				db.clear('notes');
+			} else {
+				req = db.remove('notes', model.url);
+				req.done(function () {
+					//sendResponse({msg: 'saved'});
+				});
+				req.fail(function (e) {
+					console.log(e);
+				});	
+			}
 		}
 		return req;
 	};
